Corrected CIE LAB conversion problem
For some completely unexplainable reason the B value that's calculated correctly during ColorPlus creation is correct but it's not being stored. I verified this and the only way to correct it is by storring it in yet another variable (called alt_b) and then setting it externally after object creation. I have no idea why this is happening but it now works.