lot of changes!
@withthelove

1. Menu interfaces pop up over and partially obscure the preview image.

Agreed. Added window dragging. Click anywhere that's not an input field or checkbox and drag the window where you like and it remembers the last place you place it for convenience.

2. Click 'apply' every color swap is cumbersome.

Agreed. Made the preview auto-swap colors and added an undo button in place of the apply button. It undoes color changes and removes colors added.

3. Buttons on the main window are out of order, workflow is not intuitive.

Agreed. Reordered buttons on the main window and added description text to (hopefully) make the interface more intuitive.

4. Thanks :)

#. Hex palette swap data

There's are two buttons to load/save palette swap data as a hex text file. It supports:
- hex mapping (#rrggbbaa -> #rrggbbaa)
- list of colors in hex format separated one color per line
- mixed format (hex mapping one line, single hex color next line)
- auto-map lists of colors
- comments (starting with //) at the end of a hex color / mapping line or on a separate line
- ignores extra line breaks